requirements.txt
Django==1.11.4
django-filter==1.0.4
django-material==1.0.0
djangorestframework==3.6.4
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
pytz==2017.2
